Show lead creation result as sessionId
