+++ apps/store/src/app/debugger/seb-leads/page.tsx
@@ -1,26 +1,27 @@
-import {Heading, Space, sprinkles} from "ui";
-import {SebLeadsDebuggerForm} from "@/app/debugger/seb-leads/CreateSebLeadsForm";
-import * as GridLayout from "@/components/GridLayout/GridLayout";
+import { Heading, sprinkles, Text, yStack } from 'ui'
+import { SebLeadsDebuggerForm } from '@/app/debugger/seb-leads/CreateSebLeadsForm'
+import * as GridLayout from '@/components/GridLayout/GridLayout'
 
 function SebLeadsDebuggerPage() {
-    return (
-        <Space>
-            <GridLayout.Root>
-                <GridLayout.Content width="1/3" align="center">
-                    <Space y={2}>
-                        <Heading className={sprinkles({ mt: 'xl' })} as="h1" align="center" balance={true}>
-                            Create SEB Leads
-                        </Heading>
-                        <SebLeadsDebuggerForm />
-                    </Space>
-                </GridLayout.Content>
-            </GridLayout.Root>
-        </Space>
-    )
+  if (process.env.VERCEL_ENV === 'production') {
+    return <Text>Only available in test environment</Text>
+  }
+  return (
+    <GridLayout.Root>
+      <GridLayout.Content width="1/3" align="center">
+        <div className={yStack({ gap: 'xl' })}>
+          <Heading className={sprinkles({ marginTop: 'xl' })} as="h1" align="center" balance={true}>
+            Create SEB Leads
+          </Heading>
+          <SebLeadsDebuggerForm />
+        </div>
+      </GridLayout.Content>
+    </GridLayout.Root>
+  )
 }
 
 export default SebLeadsDebuggerPage
 
 export const metadata = {
-    robots: 'noindex, nofollow',
+  robots: 'noindex, nofollow',
 }
